requirements.txt
alembic==1.6.5
amqp==5.0.6
autopep8==1.5.7
billiard==3.6.4.0
celery==5.1.2
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.4
Flask==2.0.1
Flask-API==3.0.post1
Flask-Migrate==3.0.1
Flask-SQLAlchemy==2.4.1
greenlet==1.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
kombu==5.1.0
Mako==1.1.4
MarkupSafe==2.0.1
prompt-toolkit==3.0.19
psycopg2==2.9.1
pycodestyle==2.7.0
PyJWT==2.1.0
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
pytz==2021.1
six==1.16.0
SQLAlchemy==1.4.20
toml==0.10.2
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.0.1
